# core: change block name format and add block name hints