--- a/xdsl/parser/core.py
+++ b/xdsl/parser/core.py
@@ -185,7 +185,7 @@ class Parser(AttrParser):
         if name not in self.blocks:
             self.forward_block_references[name].append(block_name)
             block = Block()
-            if Block.is_valid_name(name) and not Block._is_default_block_name(name):  # pyright: ignore[reportPrivateUsage]
+            if Block.is_valid_name(name) and not Block.is_default_block_name(name):
                 block.name_hint = name  # setter verifies validity
             self.blocks[name] = (block, None)
         return self.blocks[name][0]
@@ -252,7 +252,7 @@ class Parser(AttrParser):
             self.forward_block_references.pop(name)
 
         # Don't set name_hint for blocks that match the default pattern
-        if not Block._is_default_block_name(name):  # pyright: ignore[reportPrivateUsage]
+        if not Block.is_default_block_name(name):
             block.name_hint = name  # setter verifies validity
         # If it matches pattern "bb" followed by digits, leave name_hint as None
 
@@ -916,7 +916,7 @@ class Parser(AttrParser):
         if name not in self.blocks:
             self.forward_block_references[name].append(block_token.span)
             block = Block()
-            if not Block._is_default_block_name(name):  # pyright: ignore[reportPrivateUsage]
+            if not Block.is_default_block_name(name):
                 block.name_hint = name  # setter verifies validity
             self.blocks[name] = (block, None)
         return self.blocks[name][0]