… tests/filecheck/backend/convert_riscv_scf_to_riscv_cf_with_regalloc.mlir
… tests/filecheck/backend/riscv/prologue_epilogue_insertion.mlir
… tests/filecheck/backend/wgsl/2d5pt.mlir
… tests/filecheck/dialects/affine/affine_ops.mlir
… tests/filecheck/dialects/affine/examples.mlir
… tests/filecheck/dialects/affine/invalid.mlir
… tests/filecheck/dialects/air/air_ops.mlir
… tests/filecheck/dialects/arith/arith_cfg.mlir
… tests/filecheck/dialects/cf/canonicalize.mlir
… tests/filecheck/dialects/cf/cf_ops.mlir
… tests/filecheck/dialects/csl/csl-stencil-canonicalize.mlir
… tests/filecheck/dialects/csl/csl-stencil-ops.mlir
… tests/filecheck/dialects/csl/csl-wrapper-ops.mlir
… tests/filecheck/dialects/csl/ops.mlir
… tests/filecheck/dialects/fsm/fsm_op.mlir
… tests/filecheck/dialects/func/func_invalid.mlir
… tests/filecheck/dialects/func/func_ops.mlir
… tests/filecheck/dialects/func/func_ops_generic.mlir
… tests/filecheck/dialects/gpu/invalid.mlir
… tests/filecheck/dialects/hw/hw_module.mlir
… tests/filecheck/dialects/hw/invalid.mlir
… tests/filecheck/dialects/linalg/linalg_ops.mlir
… tests/filecheck/dialects/llvm/func.mlir
… tests/filecheck/dialects/memref_stream/canonicalize.mlir
… tests/filecheck/dialects/memref_stream/ops.mlir
… tests/filecheck/dialects/memref_stream/verify.mlir
… tests/filecheck/dialects/omp/ops.mlir
… tests/filecheck/dialects/omp/ops_invalid.mlir
… tests/filecheck/dialects/pdl_interp/pdl_interp_ops.mlir
… tests/filecheck/dialects/riscv/riscv_assembly_emission.mlir