This pr introduces name hints for blocks, similar to how an SSAValue can have a name hint. At the same time, the default format for blocks has been changed to include the `bb` prefix to match mlir-opt.