@jumerckx 2025-08-09
## Benchmarks
### Printer
* main
	```
	________________________________________________________
	Executed in   72.94 secs    fish           external
	   usr time   47.33 secs  173.00 micros   47.33 secs
	   sys time    9.22 secs  786.00 micros    9.22 secs
	```
* this pr
	```
	________________________________________________________
	Executed in   68.52 secs    fish           external
	   usr time   44.54 secs    0.15 millis   44.54 secs
	   sys time    8.48 secs    1.03 millis    8.48 secs
	```
### Parser
* main
	```
	Test Parser.constant_100 ran in: 0.00482 ± 0.000122s
	Test Parser.constant_1000 ran in: 0.0468 ± 0.00126s
	Test Parser.dense_attr ran in: 0.00899 ± 0.000437s
	Test Parser.dense_attr_hex ran in: 0.00875 ± 0.000304s
	Test Parser.large_constant_tensor ran in: 0.0864 ± 0.00445s
	
	________________________________________________________
	Executed in  160.31 secs    fish           external
	   usr time  157.16 secs    0.17 millis  157.16 secs
	   sys time    2.14 secs    1.15 millis    2.14 secs
	```
* this pr
	```
	Test Parser.constant_100 ran in: 0.00488 ± 0.000225s
	Test Parser.constant_1000 ran in: 0.0468 ± 0.00151s
	Test Parser.dense_attr ran in: 0.00896 ± 0.000243s
	Test Parser.dense_attr_hex ran in: 0.00877 ± 0.00025s
	Test Parser.large_constant_tensor ran in: 0.086 ± 0.00229s
	
	________________________________________________________
	Executed in  159.75 secs    fish           external
	   usr time  157.16 secs  137.00 micros  157.16 secs
	   sys time    1.76 secs  809.00 micros    1.76 secs
	```